added basic frontend and backend tests

+++ src/main/java/squadron/manager/turbyne/member/MemberRepository.java
@@ -3,5 +3,5 @@ package squadron.manager.turbyne.member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
-    Member findById(int id);
+    Member findById(long id);
 }
